# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trees API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://arboraid.co/beta/public"
        },
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/feed-tree/{campaignId}/donation/initiate": {
            "post": {
                "tags": [
                    "Donations"
                ],
                "summary": "Initiate donation payment via Razorpay",
                "operationId": "3c3f00c60cbe691bfaea5859c1e44071",
                "parameters": [
                    {
                        "name": "campaignId",
                        "in": "path",
                        "description": "Campaign ID for the donation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment order successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "order_id": {
                                            "type": "string",
                                            "example": "order_DjzPzU1A6Xjjtb"
                                        },
                                        "amount": {
                                            "type": "integer",
                                            "example": 50000
                                        },
                                        "currency": {
                                            "type": "string",
                                            "example": "INR"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters"
                    },
                    "404": {
                        "description": "Campaign not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/feed-tree/donation-payment/callback": {
            "post": {
                "tags": [
                    "Donations"
                ],
                "summary": "Payment callback for verifying Razorpay payment",
                "operationId": "28aca25cbc8c5624d344b90125a7b595",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "razorpay_order_id",
                                    "razorpay_payment_id",
                                    "razorpay_signature",
                                    "donation_order_id"
                                ],
                                "properties": {
                                    "donation_order_id": {
                                        "type": "string",
                                        "example": "DONOR2412180005"
                                    },
                                    "razorpay_order_id": {
                                        "type": "string",
                                        "example": "order_DjzPzU1A6Xjjtb"
                                    },
                                    "razorpay_payment_id": {
                                        "type": "string",
                                        "example": "pay_CpXjmfXZ8FgopV"
                                    },
                                    "razorpay_signature": {
                                        "type": "string",
                                        "example": "9d8b8a7d57d897a09ff3bfa6be82840bbd597f9834cb649b8f7a64b42a9cf70b5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Donation successfully processed and payment verified",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Donation successfully processed and verified."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid payment verification or missing data"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/my-donations": {
            "get": {
                "tags": [
                    "Donations"
                ],
                "summary": "Get a list of donations made by the authenticated user",
                "operationId": "7fb8f2c3676439827356ba1f604ac258",
                "responses": {
                    "200": {
                        "description": "List of user donations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "campaign_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "campaign_name": {
                                                "type": "string",
                                                "example": "Save the Forests"
                                            },
                                            "donor_name": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "amount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 500
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-12-18T12:34:56Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Enter 'Bearer {token}' to access the endpoints.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}